// product/frontend/DNAustria/src/app/api/model/eventHistoryDto.ts
/**
 * DNAustria.Api | v1
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface EventHistoryDto { 
    action: string;
    createdAt: string;
    username: string;
}